chore: Context 타입 분리 및 레퍼런스 수정
