+++ FE-MyCarMaster/src/contexts/CarPaintContext.tsx
@@ -1,43 +1,10 @@
 import { useReducer, createContext, useContext } from "react";
-
-export type ExteriorColors = {
-  id: number;
-  name: string;
-  price: number;
-  ratio: number;
-  colorImgUrl: string;
-  coloredImgUrl: string;
-};
-
-export type InteriorColors = {
-  id: number;
-  name: string;
-  price: number;
-  ratio: number;
-  colorImgUrl: string;
-  coloredImgUrl: string;
-};
-
-type CarPaintState = {
-  exteriorId: number;
-  interiorId: number;
-  exteriorList: ExteriorColors[];
-  interiorList: InteriorColors[];
-};
-
-type CarPaintAction = {
-  type:
-    | "SELECT_EXTERIOR"
-    | "SELECT_INTERIOR"
-    | "SET_EXTERIOR_LIST"
-    | "SET_INTERIOR_LIST";
-  payload: {
-    exteriorId?: number;
-    interiorId?: number;
-    exteriorList?: ExteriorColors[];
-    interiorList?: InteriorColors[];
-  };
-};
+import {
+  CarPaintState,
+  CarPaintAction,
+  ExteriorColors,
+  InteriorColors,
+} from "../types/carpaint.types";
 
 const initialCarPaintState: CarPaintState = {
   exteriorId: 1,
