+++ FE-MyCarMaster/src/types/model.types.ts
@@ -1,0 +1,9 @@
+export type ModelState = {
+  modelId: number;
+  modelName: string;
+};
+
+export type ModelAction = {
+  type: "SET_MODEL";
+  payload: { modelId: number; modelName: string };
+};
