+++ FE-MyCarMaster/src/types/quotation.types.ts
@@ -26,3 +26,32 @@ export type OptionQuotationType = {
   selectedQuotation: QuotationType[];
   consideredQuotation: QuotationType[];
 };
+
+export type QuotationState = {
+  navigationId: number;
+  isFirst: boolean[];
+  trimQuotation: TrimQuotationType;
+  detailQuotation: DetailQuotationType;
+  carPaintQuotation: CarPaintQuotationType;
+  optionQuotation: OptionQuotationType;
+};
+
+export type QuotationActionType =
+  | "NAVIGATE"
+  | "SET_TRIM_QUOTATION"
+  | "SET_DETAIL_QUOTATION"
+  | "SET_SELECT_QUOTATION"
+  | "SET_CONSIDER_QUOTATION"
+  | "SET_CAR_PAINT_QUOTATION";
+
+export type QuotationAction = {
+  type: QuotationActionType;
+  payload: {
+    id?: number;
+    navigationId?: number;
+    isFirst?: boolean[];
+    type?: string;
+    name?: string | undefined;
+    price?: number | undefined;
+  };
+};
